+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 * Replaced the `DockerContainerInitializer`'s `GetService` function with a `GetServiceFactory` function so that `NetworkContext` can manufacture `Service` instances upon `GetService` request
 * Made Golang `NetworkContext.Repartition` take in the new network state directly, rather than going through the `Repartitioner`/`RepartitionerBuilder`
+* Renamed `GetSetupTeardownBuffer` -> `GetSetupTimeout` on the `Test` interface
 
 ### Removed
 * `Repartitioner` and `RepartitionerBuilder`, which added complexity for no gain